Make version details outer layout responsive
