--- pages/version/style.scss
@@ -1,13 +1,0 @@
-@import '~@/assets/style/mixins';
-@import '~@/assets/style/dimensions';
-@import '~@/assets/style/colors';
-
-.tabAnchor {
-  display: block;
-  height: 6 * $grid;
-  padding-top: 2px;
-}
-
-.footerWrapper {
-  @include padding-y(26 * $grid);
-}
